requirements.txt
python-dotenv==1.0.0
aiogram==2.25.1
psycopg2-binary==2.9.6
